gohip-platform/server/requirements.txt
# GOHIP Platform - Backend Dependencies
# Python 3.11+ Required

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Validation & Settings
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client (for external API sourcing)
httpx>=0.26.0

# Web Scraping
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Development
python-dotenv>=1.0.0

# AI/LLM (Phase 4 - Consultant Agent)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
openai>=1.30.0  # Responses API with web_search tool support
anthropic>=0.18.0
google-generativeai>=0.4.0

# Phase 25 - Multi-Agent Orchestration
google-search-results>=2.4.2
serpapi>=0.1.5

# Phase 27 - Extended Web Research
tavily-python>=0.7.0  # Primary: 1000 free searches/month, no credit card
googlesearch-python>=1.3.0  # Fallback: Direct Google scraping
gsearch>=1.6.0  # Fallback: Alternative Google scraping

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin bcrypt for passlib compatibility
python-multipart>=0.0.6

# Encryption for API keys
cryptography>=42.0.0
